refactoring #5 - searchResultCollectionView를SearchView와 분리하기
