+++ BookSearchApp/BookSearchApp/ViewModel/SearchView/SearchResultCollectionView/HeaderView/RecentBooksCollectionViewCell.swift
