+++ BookSearchApp/BookSearchApp/ViewModel/SearchView/SearchResultCollectionView/SearchResultCollectionViewCell.swift
